+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Changelog
 Next
 ----
 
+* Add ``:date-format:`` option to the ``literalizer`` directive, allowing
+  dates and datetimes in data files to be formatted using a
+  :func:`~datetime.datetime.strftime` format string.
+
 2026.03.15.1
 ------------